+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+## 1.9.3 (2021-04-26)
+
+* Fixed missing arity on async `OutputStream` (#436)
+* Updated Jetty to 9.4.40.v20210413
+
+## 1.9.2 (2021-03-20)
+
+* Updated Jetty to 9.4.38.v20210224 (#433)
+* Fixed reflection warning (#432)
+
+## 1.9.1 (2021-02-17)
+
+* Updated Ring-Codec dependency to 1.1.3
+* Updated Jetty to 9.4.36.v20210114
+
+## 1.9.0 (2021-02-03)
+
+* Fixed automatic closing of response stream on exception (#420)
+* Changed woff media type to `font/woff` & added woff2 media type (#421)
+* Changed TTF media type to `font/ttf` (#426)
+* Improved formatting in `ring.middleware.stacktrace` (#391)
+* Improved performance of async responses (#428)
+* Added `:exclude-ciphers` option to Jetty adapter (#405)
+* Added `:exclude-protocols` option to Jetty adapter (#405)
+* Added `:ssl-context` option to Jetty adapter (#412)
+* Added `:async-timeout-handler` option to Jetty adapter (#410)
+* Added `:keystore-scan-interval` option to Jetty adapter (#417)
+
+## 1.8.2 (2020-10-06)
+
+* Updated Jetty to 9.4.31.v20200723 (#411)
+
+## 1.8.1 (2020-05-02)
+
+* Deprecated string secret keys for cookie session stores
+* Fixed `nil` bug in `wrap-resource` (#385)
+* Fixed charset parsing not reading quoted values (#402)
+* Fixed missing type hint in Jetty adapter (#401)
+* Updated Jetty to 9.4.28.v2020040
+* Added ring-bench benchmarking subproject
+
+## 1.8.0 (2019-11-13)
+
+* Fixed confusing exception on bad queue configuration (#354)
+* Fixed poor content-type for `wrap-stacktrace` (#378)
+* Fixed client cert support broken by deprecated class (#380)
+* Fixed namespace reload order in `wrap-reload` (#377)
+* Fixed second call of handler on exceptions (#365)
+* Improved performance of `ring.request/content-type` (#332)
+* Updated Jetty to 9.4.22.v20191022
+* Updated Commons-FileUpload to 1.4
+* Updated Ring-Codec to 1.1.2
+* Updated ns-tracker to 0.4.0
+* Added `:none` to `Same-Site` cookie header (#374)
+* Added server connector type hint for GraalVM support (#381)
+* Added support for byte array response bodies (#334)
+* Removed clj-time as a mandatory dependency (#359)
+
 ## 1.7.1 (2018-10-27)
 
 * Updated Jetty to 9.4.12.v20180830